fix: moved `addFraxSeries` & `updateDust`
