+++ scripts/governance/add/addSeries/addFraxSeries/addFraxSeries.goerli.config.ts
@@ -30,10 +30,10 @@ import {
   FYETH2209,
   FYUSDC2206,
   FYUSDC2209,
-} from '../../../../shared/constants'
-import { readAddressMappingIfExists } from '../../../../shared/helpers'
+} from '../../../../../shared/constants'
+import { readAddressMappingIfExists } from '../../../../../shared/helpers'
 
-import * as base_config from '../../base.kovan.config'
+import * as base_config from '../../../base.goerli.config'
 
 export const chainId: number = base_config.chainId
 export const developer: string = '0x5AD7799f02D5a829B2d6FA085e6bd69A872619D5'
@@ -43,10 +43,10 @@ export const whales: Map<string, string> = basewhales.set(FRAX, developer)
 export const governance: Map<string, string> = base_config.governance
 export const protocol: Map<string, string> = base_config.protocol
 export const assets: Map<string, string> = base_config.assets
-// export const joins: Map<string, string> = base_config.joins
-// export const newFYTokens: Map<string, string> = base_config.newFYTokens
-// export const newPools: Map<string, string> = base_config.newPools
-// export const newStrategies: Map<string, string> = base_config.newStrategies
+export const joins: Map<string, string> = base_config.joins
+export const newFYTokens: Map<string, string> = base_config.newFYTokens
+export const newPools: Map<string, string> = base_config.newPools
+export const newStrategies: Map<string, string> = base_config.newStrategies
 export const newJoins: Map<string, string> = readAddressMappingIfExists('newJoins.json')
 
 export const rateChiSources: Array<[string, string, string, string]> = [
@@ -81,7 +81,7 @@ export const newCompositePaths: Array<[string, string, Array<string>]> = [
 
 // Input data: baseId, ilkId, ratio (1000000 == 100%), line, dust, dec
 export const newChainlinkLimits: Array<[string, string, number, number, number, number]> = [
-  [FRAX, FRAX, 1000000, 5000000, 5000, 18],
+  [FRAX, FRAX, 1000000, 5000000, 0, 18],
   [FRAX, ETH, 1400000, 1000000, 5000, 18],
   [FRAX, DAI, 1100000, 1000000, 5000, 18],
   [FRAX, USDC, 1100000, 1000000, 5000, 18],
