+++ scripts/governance/add/addSeries/addFraxSeries/addFraxSeries.mainnet.config.ts
@@ -30,10 +30,10 @@ import {
   FYDAI2209,
   FYETH2209,
   FYUSDC2209,
-} from '../../../../shared/constants'
-import { readAddressMappingIfExists } from '../../../../shared/helpers'
+} from '../../../../../shared/constants'
+import { readAddressMappingIfExists } from '../../../../../shared/helpers'
 
-import * as base_config from '../../base.mainnet.config'
+import * as base_config from '../../../base.mainnet.config'
 
 export const chainId: number = base_config.chainId
 export const developer: string = '0xC7aE076086623ecEA2450e364C838916a043F9a8'
