+++ scripts/governance/add/addSeries/addFraxSeries/addFraxSeries.test.ts
@@ -8,9 +8,9 @@ import {
   impersonate,
   getOwnerOrImpersonate,
   getOriginalChainId,
-} from '../../../../shared/helpers'
-import { ERC20Mock, Cauldron, Ladle, FYToken, CompositeMultiOracle, ChainlinkMultiOracle } from '../../../../typechain'
-import { ENS, FRAX, FYETH2206, FYETH2209, UNI, WAD, WSTETH } from '../../../../shared/constants'
+} from '../../../../../shared/helpers'
+import { ERC20Mock, Cauldron, Ladle, FYToken, CompositeMultiOracle, ChainlinkMultiOracle } from '../../../../../typechain'
+import { ENS, FRAX, FYETH2206, FYETH2209, UNI, WAD, WSTETH } from '../../../../../shared/constants'
 const { developer, seriesIlks, assets, whales } = require(process.env.CONF as string)
 
 /**
