+++ scripts/governance/add/addSeries/addFraxSeries/mocks/deployFraxAggregatorMock.ts
@@ -1,9 +1,9 @@
 import { ethers, waffle } from 'hardhat'
-import { ETH, FRAX } from '../../../../../shared/constants'
-import { verify, getOriginalChainId, readAddressMappingIfExists, writeAddressMap } from '../../../../../shared/helpers'
+import { ETH, FRAX } from '../../../../../../shared/constants'
+import { verify, getOriginalChainId, readAddressMappingIfExists, writeAddressMap } from '../../../../../../shared/helpers'
 
-import ChainlinkAggregatorV3MockArtifact from '../../../../../artifacts/contracts/::mocks/ChainlinkAggregatorV3Mock.sol/ChainlinkAggregatorV3Mock.json'
-import { ChainlinkAggregatorV3Mock } from '../../../../../typechain'
+import ChainlinkAggregatorV3MockArtifact from '../../../../../../artifacts/contracts/::mocks/ChainlinkAggregatorV3Mock.sol/ChainlinkAggregatorV3Mock.json'
+import { ChainlinkAggregatorV3Mock } from '../../../../../../typechain'
 
 const { deployContract } = waffle
