+++ scripts/governance/add/addSeries/addFraxSeries/mocks/deployFraxMock.ts
@@ -5,10 +5,10 @@ import {
   readAddressMappingIfExists,
   writeAddressMap,
   getOwnerOrImpersonate,
-} from '../../../../../shared/helpers'
+} from '../../../../../../shared/helpers'
 
-import ERC20MockArtifact from '../../../../../artifacts/contracts/::mocks/ERC20Mock.sol/ERC20Mock.json'
-import { ERC20Mock } from '../../../../../typechain/ERC20Mock'
+import ERC20MockArtifact from '../../../../../../artifacts/contracts/::mocks/ERC20Mock.sol/ERC20Mock.json'
+import { ERC20Mock } from '../../../../../../typechain/ERC20Mock'
 
 const { deployContract } = waffle
 const { developer } = require(process.env.CONF as string)
