+++ scripts/governance/update/updateDust/updateDust.ts
@@ -11,9 +11,9 @@ import {
   getOriginalChainId,
   proposeApproveExecute,
   readAddressMappingIfExists,
-} from '../../../shared/helpers'
-import { updateDustProposal } from '../../fragments/limits/updateDustProposal'
-import { Cauldron, Timelock } from '../../../typechain'
+} from '../../../../shared/helpers'
+import { updateDustProposal } from '../../../fragments/limits/updateDustProposal'
+import { Cauldron, Timelock } from '../../../../typechain'
 const { governance, protocol, developer, newMin } = require(process.env.CONF as string)
 ;(async () => {
   let ownerAcc = await getOwnerOrImpersonate(developer)
